Fix more IDisposables and nulls

+++ WalletWasabi.Packager/MacSignTools.cs
@@ -53,10 +53,21 @@ namespace WalletWasabi.Packager
 
 			Console.WriteLine("Phase: creating the working directory.");
 
-			Console.WriteLine("Enter appleId (email):");
-			var appleId = Console.ReadLine();
-			Console.WriteLine("Enter password:");
-			var password = Console.ReadLine();
+			string? appleId = null;
+			string? password = null;
+			do
+			{
+				Console.WriteLine("Enter appleId (email):");
+				appleId = Console.ReadLine();
+			}
+			while (string.IsNullOrWhiteSpace(appleId));
+
+			do
+			{
+				Console.WriteLine("Enter password:");
+				password = Console.ReadLine();
+			}
+			while (string.IsNullOrWhiteSpace(password));
 
 			if (Directory.Exists(workingDir))
 			{
@@ -80,7 +91,7 @@ namespace WalletWasabi.Packager
 			Console.WriteLine("Update the plist file with current information for example with version.");
 
 			var lines = File.ReadAllLines(infoFilePath);
-			string bundleIdentifier = null;
+			string? bundleIdentifier = null;
 
 			for (int i = 0; i < lines.Length; i++)
 			{
@@ -100,6 +111,10 @@ namespace WalletWasabi.Packager
 					bundleIdentifier = lines[i + 1].Trim().Replace("<string>", "").Replace("</string>", "");
 				}
 			}
+			if (string.IsNullOrWhiteSpace(bundleIdentifier))
+			{
+				throw new InvalidDataException("Bundle identifier not found in plist file.");
+			}
 
 			File.Delete(infoFilePath);
 
@@ -112,7 +127,7 @@ namespace WalletWasabi.Packager
 				WorkingDirectory = workingDir
 			}))
 			{
-				process.WaitForExit();
+				WaitProcessToFinish(process, "chmod");
 			}
 
 			var filesToCheck = new[] { entitlementsPath };
@@ -146,7 +161,7 @@ namespace WalletWasabi.Packager
 					Arguments = $"u+x \"{file}\"",
 					WorkingDirectory = workingDir
 				});
-				process.WaitForExit();
+				WaitProcessToFinish(process, "chmod");
 			}
 
 			SignDirectory(filesToSignInOrder, workingDir, signArguments, entitlementsPath);
@@ -165,7 +180,7 @@ namespace WalletWasabi.Packager
 				WorkingDirectory = workingDir
 			}))
 			{
-				process.WaitForExit();
+				WaitProcessToFinish(process, "ditto");
 			}
 
 			Notarize(appleId, password, appNotarizeFilePath, bundleIdentifier);
@@ -179,8 +194,8 @@ namespace WalletWasabi.Packager
 				RedirectStandardError = true
 			}))
 			{
-				process.WaitForExit();
-				string result = process.StandardError.ReadToEnd();
+				var nonNullProcess = WaitProcessToFinish(process, "spctl");
+				string result = nonNullProcess.StandardError.ReadToEnd();
 				if (!result.Contains(": accepted"))
 				{
 					throw new InvalidOperationException(result);
@@ -210,7 +225,7 @@ namespace WalletWasabi.Packager
 				WorkingDirectory = dmgPath
 			}))
 			{
-				process.WaitForExit();
+				WaitProcessToFinish(process, "ln");
 			}
 
 			var hdutilCreateArgs = string.Join(
@@ -232,7 +247,7 @@ namespace WalletWasabi.Packager
 				WorkingDirectory = dmgPath
 			}))
 			{
-				process.WaitForExit();
+				WaitProcessToFinish(process, "hdiutil");
 			}
 
 			var hdutilConvertArgs = string.Join(
@@ -252,7 +267,7 @@ namespace WalletWasabi.Packager
 				WorkingDirectory = dmgPath
 			}))
 			{
-				process.WaitForExit();
+				WaitProcessToFinish(process, "hdiutil");
 			}
 
 			Console.WriteLine("Phase: signing the dmg file.");
@@ -277,8 +292,8 @@ namespace WalletWasabi.Packager
 				RedirectStandardError = true
 			}))
 			{
-				process.WaitForExit();
-				string result = process.StandardError.ReadToEnd();
+				var nonNullProcess = WaitProcessToFinish(process, "spctl");
+				string result = nonNullProcess.StandardError.ReadToEnd();
 				if (!result.Contains(": accepted"))
 				{
 					throw new InvalidOperationException(result);
@@ -291,9 +306,19 @@ namespace WalletWasabi.Packager
 			Console.WriteLine("Phase: finish.");
 		}
 
+		private static Process WaitProcessToFinish(Process? process, string processName)
+		{
+			if (process is null)
+			{
+				throw new InvalidOperationException($"Could not start ${processName} process.");
+			}
+			process.WaitForExit();
+			return process;
+		}
+
 		private static void Notarize(string appleId, string password, string filePath, string bundleIdentifier)
 		{
-			string uploadId = null;
+			string? uploadId = null;
 
 			Console.WriteLine("Start notarizing, uploading file.");
 
@@ -304,8 +329,8 @@ namespace WalletWasabi.Packager
 				RedirectStandardOutput = true,
 			}))
 			{
-				process.WaitForExit();
-				string result = process.StandardOutput.ReadToEnd();
+				var nonNullProcess = WaitProcessToFinish(process, "xcrum");
+				string result = nonNullProcess.StandardOutput.ReadToEnd();
 
 				if (result.Contains("The software asset has already been uploaded. The upload ID is"))
 				{
@@ -351,8 +376,8 @@ namespace WalletWasabi.Packager
 					RedirectStandardError = true,
 					RedirectStandardOutput = true,
 				});
-				process.WaitForExit();
-				string result = $"{process.StandardError.ReadToEnd()} {process.StandardOutput.ReadToEnd()}";
+				var nonNullProcess = WaitProcessToFinish(process, "xcrum");
+				string result = $"{nonNullProcess.StandardError.ReadToEnd()} {nonNullProcess.StandardOutput.ReadToEnd()}";
 				if (result.Contains("Status Message: Package Approved"))
 				{
 					break;
@@ -379,7 +404,7 @@ namespace WalletWasabi.Packager
 				FileName = "xcrun",
 				Arguments = $"stapler staple \"{filePath}\"",
 			});
-			process.WaitForExit();
+			WaitProcessToFinish(process, "xcrum");
 		}
 
 		private static void DeleteWithChmod(string path)
@@ -390,7 +415,7 @@ namespace WalletWasabi.Packager
 				Arguments = $"-R ugo+rwx \"{path}\"",
 			}))
 			{
-				process.WaitForExit();
+				WaitProcessToFinish(process, "chmod");
 			}
 
 			IoHelpers.TryDeleteDirectoryAsync(path).GetAwaiter().GetResult();
@@ -405,8 +430,8 @@ namespace WalletWasabi.Packager
 				WorkingDirectory = workingDir,
 				RedirectStandardError = true
 			});
-			process.WaitForExit();
-			var result = process.StandardError.ReadToEnd();
+			var nonNullProcess = WaitProcessToFinish(process, "codesign");
+			var result = nonNullProcess.StandardError.ReadToEnd();
 			if (result.Contains("code object is not signed at all"))
 			{
 				throw new InvalidOperationException(result);
@@ -428,8 +453,8 @@ namespace WalletWasabi.Packager
 				Arguments = $"-dv --verbose=4 \"{path}\"",
 				RedirectStandardError = true,
 			});
-			process.WaitForExit();
-			string result = process.StandardError.ReadToEnd();
+			var nonNullProcess = WaitProcessToFinish(process, "codesign");
+			string result = nonNullProcess.StandardError.ReadToEnd();
 			if (!result.Contains("Authority=Developer ID Application: zkSNACKs Ltd."))
 			{
 				throw new InvalidOperationException(result);
@@ -480,7 +505,10 @@ namespace WalletWasabi.Packager
 				RedirectStandardOutput = true,
 				CreateNoWindow = true
 			});
-
+			if (process is null)
+			{
+				throw new InvalidOperationException("Could not start bash process.");
+			}
 			var result = process.StandardOutput.ReadToEnd();
 			process.WaitForExit();
